requirements.txt
django
djangorestframework
django-cors-headers
python-dotenv
langchain
openai
faiss-cpu
langchain-community